fix(preprocessing): ensure segment data type consistency in TUEVEventExtractor
- Updated the segment data type in the TUEVEventExtractor to explicitly convert segments to float32 and annotations to integers before appending to the segments list.
- This change enhances data type consistency and ensures compatibility with downstream processing, reinforcing adherence to expected input formats.
- Maintained functionality while improving the robustness of the segment extraction process.

+++ src/brain_go_brrr/infra/preprocessing/tuev_event_extractor.py
@@ -141,6 +141,6 @@ class TUEVEventExtractor:
                 f"Segment shape {segment.shape} != (23, {samples_per_segment})"
             )
 
-            segments.append((segment, annot['label']))
+            segments.append((segment.astype(np.float32), int(annot['label'])))
 
         return segments
